Added a ENABLE_DEV_MODE option to turn on developer mode.  By default this mode is off.
In developer mode, all warnings are turned on, the sanitizers are
turned on and verbose output is enabled. The build config is also set
to Debug to enable debugging.

When developer mode is NOT enabled (the default), the build config is
set to Release for the fastest code and verbose output is shut off to
make the build look nicer.

In this setup, you can have two build folders and configure them
differently so you can switch back and forth easily.

mkdir buildRelease
cd buildRelease
cmake ..

versus

mkdir buildDebug
cd buildDebug
cmake -DENABLE_DEV_MODE=ON ..